+++ spit/classify.py
@@ -90,3 +90,23 @@ def classify_me(image_file, classifier_root=None, verbose=False):
 
     # Return
     return il.Frames(prediction).name
+
+
+def one_hot_encoded(class_numbers, num_classes=None):
+    """
+    Generate the One-Hot encoded class-labels from an array of integers.
+    For example, if class_number=2 and num_classes=4 then
+    the one-hot encoded label is the float array: [0. 0. 1. 0.]
+    :param class_numbers:
+        Array of integers with class-numbers.
+        Assume the integers are from zero to num_classes-1 inclusive.
+    :param num_classes:
+        Number of classes. If None then use max(cls)-1.
+    :return:
+        2-dim array of shape: [len(cls), num_classes]
+    """
+    # Find the number of classes if None is provided.
+    if num_classes is None:
+        num_classes = np.max(class_numbers) - 1
+
+    return np.eye(num_classes, dtype=float)[class_numbers]
