+++ spit/preprocess.py
@@ -1,59 +1,64 @@
+""" Pre-procesing methods for an input image
+"""
 from __future__ import print_function, absolute_import, division, unicode_literals
 
-import numpy as np, glob, math
-from astropy.io import fits
-from scipy import misc
+import numpy as np
+from PIL import Image
 import tensorflow as tf
 import pdb
 
 from spit import zscale as aut_z
-
-########################################################################
-# Various constants for the size of the images.
-# Use these constants in your own program.
-
-# Width and height of each image.
-# The height of an image
-image_height = 210
-
-# The width of an image
-image_width = 650
-
-# Length of an image when flattened to a 1-dim array.
-img_size_flat = image_height * image_width
-
-# Tuple with height and width of images used to reshape arrays.
-img_shape = (image_height, image_width)
-
-# Number of channels in each image, 3 channels: Red, Green, Blue.
-num_channels = 1
-
-# Number of classes.
-num_classes = 5
-
-# The padding value for the padded image
-pad_const = 0
-########################################################################
+from spit import defs
 
 
-def one_hot_encoded(class_numbers, num_classes=None):
+def process_image(image, flips=True):
+    """ Process a single image
+    
+    Parameters
+    ----------
+    image : ndarray
+
+    Returns
+    -------
+    zimage : ndarray,  if flips=False
+       2D trimmed, resized, scaled image
+    images_array : ndarray,  if flips=True
+       Flattened, 4 flip version
     """
-    Generate the One-Hot encoded class-labels from an array of integers.
-    For example, if class_number=2 and num_classes=4 then
-    the one-hot encoded label is the float array: [0. 0. 1. 0.]
-    :param class_numbers:
-        Array of integers with class-numbers.
-        Assume the integers are from zero to num_classes-1 inclusive.
-    :param num_classes:
-        Number of classes. If None then use max(cls)-1.
-    :return:
-        2-dim array of shape: [len(cls), num_classes]
-    """
-    # Find the number of classes if None is provided.
-    if num_classes is None:
-        num_classes = np.max(class_numbers) - 1
+    from spit.utils import congrid
+    # Trim
+    image = trim_image(image, cutoff_percent=defs.cutoff_percent)
 
-    return np.eye(num_classes, dtype=float)[class_numbers]
+    # Resize
+    rimage = congrid(image.astype(float), (defs.image_height, defs.image_width))
+
+    # zscale
+    zimage = zscale(rimage)
+
+    # Flips?
+    if flips:
+        pil_image = Image.fromarray(zimage)
+
+        # Generate flipped images
+        pil_image.transpose(Image.FLIP_TOP_BOTTOM)
+        ver_image = np.array(pil_image.transpose(Image.FLIP_TOP_BOTTOM))
+        hor_image = np.array(pil_image.transpose(Image.FLIP_LEFT_RIGHT))
+        hor_ver_image = np.array(pil_image.transpose(Image.FLIP_LEFT_RIGHT).transpose(Image.FLIP_TOP_BOTTOM))
+
+        # Add flipped images to an array
+        im_array = []
+        im_array.append(zimage.flatten())
+        im_array.append(ver_image.flatten())
+        im_array.append(hor_image.flatten())
+        im_array.append(hor_ver_image.flatten())
+        images_array = np.array(im_array)
+
+        # Return
+        return images_array
+    else:
+        # Return
+        return zimage
+
 
 def resize_image(image_data, sess):
     # Convert the image data to an int
@@ -68,127 +73,27 @@ def resize_image(image_data, sess):
     image_data = image_data[..., np.newaxis]
 
     # Define the new sizes keeping the aspect ration
-    new_height = int(image_width / aspect_ratio)
+    new_height = int(defs.image_width / aspect_ratio)
 
     # User tensorflow to resize the image
-    new_image = tf.image.resize_images(image_data, [new_height, image_width])
+    new_image = tf.image.resize_images(image_data, [new_height, defs.image_width])
     new_image = new_image.eval(session=sess)
     
     # Remove the third added dimension
     new_image = new_image[:, :, 0]
 
     # Add padding to make the image 200x600
-    height_pad = image_height - new_image.shape[0]
-    width_pad = image_width - new_image.shape[1]
+    height_pad = defs.image_height - new_image.shape[0]
+    width_pad = defs.image_width - new_image.shape[1]
     npad = ((0, height_pad), (0, width_pad))
     
-    padded_image = np.pad(new_image, pad_width=npad, mode='constant', constant_values=pad_const) 
+    padded_image = np.pad(new_image, pad_width=npad, mode='constant', constant_values=defs.pad_const)
     
     return padded_image
-
-def load_data_cropped200x600(data_type, img_path='/soe/vjankov/scratchdisk/'):
-    image_data = {}
-    """
-    a)
-    Load the data from the filenames stored in the data_locations 
-
-    b)
-    Add 0 padding to each image. The padding is 2112 pixels width and height
-    The 2112 is the widest pixel form the dataset
-    
-    c)
-    Add labels for each image based on the folder they are coming form
-    This is the label names and values
-
-    Bias = 0
-    Science = 1
-    Arc = 2
-    Flat = 3
-    
-    d)
-    Construct a dictionary with the following properties:
-        raw_data = the flattened 2112 np array with raw pixel values
-        labels = the corresponding labels for each data element
-        filename = the corresponding filename
-
-    """
-    # Define the image locations
-    if data_type == "train_data":
-        load_batch_size = 5400
-
-        data_locations = [ \
-                img_path+"viktor_astroimage/bias_train/*", \
-                img_path+"viktor_astroimage/science_train/*", \
-                img_path+"viktor_astroimage/standard_train/*", \
-                img_path+"viktor_astroimage/arc_train/*", \
-                img_path+"viktor_astroimage/flat_train/*"]
-        
-    elif data_type == "test_data":
-        load_batch_size = 1650
-
-        data_locations = [ \
-                img_path+"viktor_astroimage/bias_test/*", \
-                img_path+"viktor_astroimage/science_test/*", \
-                img_path+"viktor_astroimage/standard_test/*", \
-                img_path+"viktor_astroimage/arc_test/*", \
-                img_path+"viktor_astroimage/flat_test/*"]
-
-    elif data_type == "validation_data":
-        load_batch_size = 1350
-        data_locations = [ \
-                img_path+"viktor_astroimage/bias_validation/*", \
-                img_path+"viktor_astroimage/science_validation/*", \
-                img_path+"viktor_astroimage/standard_validation/*", \
-                img_path+"viktor_astroimage/arc_validation/*", \
-                img_path+"viktor_astroimage/flat_validation/*"]
-        
-    # Construct the dict arrays
-    raw_data = []
-    labels = []
-    filenames = []
-
-    for index, location in enumerate(data_locations):
-        images = glob.glob(location)
-        image_array = []
-        image_labels = []
-        image_filenames = []
-        for image_file in images:
-            hdulist = fits.open(image_file)
-            image_data = hdulist[0].data
-            padded_image = image_data.flatten()
-
-            image_array.append(padded_image)
-            image_labels.append(index)
-            image_filenames.append(image_file)
-
-        """
-        while len(image_array) < load_batch_size:
-            image_array = image_array + image_array
-            image_labels = image_labels + image_labels
-            image_filenames = image_filenames + image_filenames
-
-        raw_data = raw_data + image_array[:load_batch_size]
-        labels = labels + image_labels[:load_batch_size]
-        filenames = filenames + image_filenames[:load_batch_size]
-        """
-
-        raw_data = raw_data + image_array
-        labels = labels + image_labels
-        filenames = filenames + image_filenames
-
-    # Get the raw images.
-    raw_images = np.array(raw_data)
-    
-    # Get the class-numbers for each image. Convert to numpy-array.
-    cls = np.array(labels)
-
-    return raw_images, cls, one_hot_encoded(class_numbers=cls, num_classes=num_classes), filenames
 
 
 def cutoff_forw(max_vals, cutoff_percent=1.15):
     cutoff_point = 0
-    prev_val = 0
-    curr_val = 0
     for idx in range(1, len(max_vals)):
         prev_val = max_vals[idx - 1]
         curr_val = max_vals[idx]
@@ -204,8 +109,6 @@ def cutoff_back(max_vals, cutoff_percent=1.15):
     """ Trim image from the top"""
     max_vals.reverse()
     cutoff_point = 0
-    prev_val = 0
-    curr_val = 0
     for idx in range(1, len(max_vals)):
         prev_val = max_vals[idx - 1]
         curr_val = max_vals[idx]
